Java_Lessons-185 101: Swing 6: WindowListener 185 https://youtu.be/GPmJPNq38Yk?t=3m37s WindowListener - следит за окном. Типа : "Вы уверены? Да/Нет" и др. Хотя можно выполнять любой код Очень нечитабельно.
